refactor: [US-049] - Move profile-update views into core/views/profile.py
