+++ core/views/profile.py
@@ -1,0 +1,135 @@
+"""Profile settings views: privacy, display name, and recommendation-visibility updates."""
+
+import json
+import logging
+
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+from django.shortcuts import redirect
+from django.template.loader import render_to_string
+from django.urls import reverse
+from django.views.decorators.http import require_POST
+from django_ratelimit.decorators import ratelimit
+from django_ratelimit.exceptions import Ratelimited
+
+from ..analytics.events import track_profile_made_public, track_settings_updated
+from ..cache_utils import safe_cache_delete
+from ..forms import UpdateDisplayNameForm
+
+logger = logging.getLogger(__name__)
+
+
+@login_required
+@require_POST
+def update_privacy_view(request):
+    is_public = request.POST.get("is_public") == "true"
+    profile = request.user.userprofile
+    profile.is_public = is_public
+    profile.save()
+
+    if is_public:
+        # Track profile made public
+        track_profile_made_public(request.user.id)
+
+        public_url = request.build_absolute_uri(
+            reverse("core:public_profile", kwargs={"username": request.user.username})
+        )
+        message_text = render_to_string(
+            "core/partials/messages_with_link.html",
+            {"public_url": public_url, "username": request.user.username},
+        )
+        messages.success(request, message_text)
+    else:
+        messages.success(request, "Your profile is now private.")
+
+    if "dna_data" in request.session:
+        request.session.pop("dna_data", None)
+
+    return redirect("core:display_dna")
+
+
+@login_required
+@require_POST
+def update_display_name_view(request):
+    form = UpdateDisplayNameForm(request.POST, user=request.user, instance=request.user)
+    if form.is_valid():
+        form.save()
+        # Track settings update
+        track_settings_updated(request.user.id, setting_type="display_name")
+        messages.success(request, "Your display name has been updated!")
+    else:
+        for error in form.errors.values():
+            messages.error(request, error)
+
+    return redirect("core:display_dna")
+
+
+@ratelimit(key="user", rate="10/m", method="POST", block=True)
+def _update_username_api_throttled(request):
+    try:
+        data = json.loads(request.body)
+        new_username = data.get("username")
+
+        if not new_username:
+            return JsonResponse({"status": "error", "message": "Display name cannot be empty."}, status=400)
+
+        form = UpdateDisplayNameForm({"username": new_username}, user=request.user, instance=request.user)
+
+        if form.is_valid():
+            form.save()
+            messages.success(request, "Display name updated successfully!")
+
+            return JsonResponse({"status": "success", "new_username": new_username})
+        else:
+            error_message = form.errors.get("username")[0]
+            return JsonResponse({"status": "error", "message": error_message}, status=400)
+
+    except Exception as e:
+        logger.error(f"Error in update_username_api: {e}", exc_info=True)
+
+        return JsonResponse({"status": "error", "message": "An unexpected server error occurred."}, status=500)
+
+
+@login_required
+@require_POST
+def update_username_api(request):
+    try:
+        return _update_username_api_throttled(request)
+    except Ratelimited:
+        return JsonResponse({"error": "Too many attempts, try again later."}, status=429)
+
+
+@login_required
+@require_POST
+def update_recommendation_visibility(request):
+    """Toggle visibility in recommendations"""
+    is_visible = request.POST.get("visible_in_recommendations") == "true"
+    profile = request.user.userprofile
+    was_visible = profile.visible_in_recommendations
+    profile.visible_in_recommendations = is_visible
+    profile.save()
+
+    # Invalidate caches keyed on this user's recommendation/similarity output so
+    # the toggle takes effect immediately (US-024c).
+    safe_cache_delete(f"user_recommendations_{request.user.id}")
+    safe_cache_delete(f"similar_users_{request.user.id}")
+
+    # When opting OUT, also flush the candidate-pool cache so the user is
+    # dropped from other readers' similarity searches on the next refresh.
+    if was_visible and not is_visible:
+        safe_cache_delete("public_users_for_recs_sample")
+        logger.info(
+            "user opted out of recs; cleared candidate sample cache",
+            extra={"user_id": request.user.id},
+        )
+
+    # Track settings update
+    track_settings_updated(request.user.id, setting_type="recommendation_visibility")
+
+    if is_visible:
+        messages.success(request, "You are now visible as a recommendation source to similar readers!")
+    else:
+        messages.success(request, "You've opted out of being shown as a recommendation source.")
+
+    return redirect("core:display_dna")
